Add meta tag with UTF-8 charset to load Fabric script properly
The absence of meta tag is probably tampering the encoding or may be
Fabric.js is in a different encoding, I am not so sure but, during
debugging the presence of some weird characters made me enforce the
charset to UTF-8 explicitly on the GettingStarted.html.

This is just my theory, I could be completely wrong. Please refer to
the actual documentation for better understanding.